+++ src/View/Helper/TaxonomyTermSelect.php
@@ -1,0 +1,29 @@
+<?php
+namespace Taxonomy\View\Helper;
+
+use Taxonomy\Form\Element\TaxonomyTermSelect as Select;
+use Laminas\Form\Factory;
+use Laminas\View\Helper\AbstractHelper;
+use Laminas\ServiceManager\ServiceLocatorInterface;
+
+class TaxonomyTermSelect extends AbstractHelper
+{
+    protected $formElementManager;
+
+    public function __construct(ServiceLocatorInterface $formElementManager)
+    {
+        $this->formElementManager = $formElementManager;
+    }
+
+    public function __invoke(array $spec = [])
+    {
+        $spec['type'] = Select::class;
+        if (!isset($spec['options']['empty_option'])) {
+            $spec['options']['empty_option'] = 'Select taxonomy term…'; // @translate
+        }
+        $factory = new Factory($this->formElementManager);
+        $element = $factory->createElement($spec);
+
+        return $this->getView()->formTaxonomyTermSelect($element);
+    }
+}